// backend/src/types/database.types.ts
export interface PrismaClient {
  organization: any;
  user: any;
  product: any;
  contentGeneration: any;
  scheduledContent: any;
  bulkOperation: any;
  webhookLog: any;
  auditLog: any;
  businessProfile: any;
  customQuestionTemplate: any;
  qaPage: any;
  internalLink: any;
  contentPerformance: any;
  abTest: any;
  automationRule: any;
  competitor: any;
  $transaction: (queries: any[]) => Promise<any>;
  $disconnect: () => Promise<void>;
}